requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
atlastk==0.13.5
certifi==2025.6.15
ciso8601==2.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
orjson==3.10.18
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.1
python-dotenv==1.1.0
pytz==2025.2
sniffio==1.3.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvloop==0.21.0
websockets==15.0.1
